requirements.txt
flask
google-api-python-client
python-gflags
google-apputils